Created AppLoading widget Added constants file
