+++ src/app/AppLoading/index.ts
@@ -1,0 +1,1 @@
+export { default as AppLoading } from './AppLoadingWidget';
